Test correct detects the problem
